refactor: renaming and move package - Enum files
