+++ src/main/java/com/example/demo/domain/mission/MissionCategoryEnum.java
@@ -1,6 +1,6 @@
-package com.example.demo.domain.common;
+package com.example.demo.domain.mission;
 
-public enum Category {
+public enum MissionCategoryEnum {
     REFRESH,        // 리프레시
     EMPLOYMENT,     // 취업
     DAILY           // 일상
